+++ src/lib.rs
@@ -125,13 +125,6 @@
 //!
 //! See [auto_enums crate](https://github.com/taiki-e/auto_enums/tree/master/derive/src/derive) for more examples.
 //!
-//! ## Crate Features
-//!
-//! * `std`
-//!   * Enabled by default.
-//!   * Generate code for `std` library.
-//!   * Disable this feature to generate code for `no_std`.
-//!
 //! ## Rust Version
 //!
 //! The current minimum required Rust version is 1.30.
@@ -152,7 +145,6 @@ mod common;
 mod error;
 mod parse;
 
-pub use self::common::std_root;
 pub use self::error::{Error, Result, *};
 pub use self::parse::*;