+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+* Remove `std` feature and `std_root` function<br>
+  `derive_utils` can generate accurate code without `std` feature.
+
 # 0.5.4 - 2019-01-03
 
 * Improve documentation